+++ src/Database/Beam/MySQL/FromField.hs
@@ -1,10 +1,7 @@
-{-# LANGUAGE AllowAmbiguousTypes    #-}
-{-# LANGUAGE FunctionalDependencies #-}
-{-# LANGUAGE GADTs                  #-}
-{-# LANGUAGE KindSignatures         #-}
-{-# LANGUAGE ScopedTypeVariables    #-}
-{-# LANGUAGE StandaloneDeriving     #-}
-{-# LANGUAGE UndecidableInstances   #-}
+{-# LANGUAGE AllowAmbiguousTypes  #-}
+{-# LANGUAGE ScopedTypeVariables  #-}
+{-# LANGUAGE TypeFamilies         #-}
+{-# LANGUAGE UndecidableInstances #-}
 
 module Database.Beam.MySQL.FromField where
 
@@ -20,449 +17,303 @@ import           Data.Word (Word16, Word32, Word64, Word8)
 import           Database.Beam.Backend.SQL (SqlNull (SqlNull))
 import           Database.MySQL.Base (MySQLValue (..))
 import           Text.Read (readMaybe)
-import           Type.Reflection (TyCon, Typeable, tyConName, typeRep,
-                                  typeRepTyCon)
+import           Type.Reflection (TyCon, Typeable, typeRep, typeRepTyCon)
 
-data Leniency = Lenient | Strict
+data Strict =
+  UnexpectedNull |
+  TypeMismatch |
+  Won'tFit
   deriving stock (Eq, Show)
 
-data FromFieldResult (l :: Leniency) (a :: Type) where
-  UnexpectedNull' :: FromFieldResult l a
-  TypeMismatch' :: FromFieldResult l a
-  Won'tFit' :: FromFieldResult l a
-  IEEENaN' :: FromFieldResult 'Lenient a
-  IEEEInfinity' :: FromFieldResult 'Lenient a
-  IEEETooSmall' :: FromFieldResult 'Lenient a
-  IEEETooBig' :: FromFieldResult 'Lenient a
-  TextCouldNotParse' :: FromFieldResult 'Lenient a
-  LenientParse :: a -> FromFieldResult 'Lenient a
-  StrictParse :: a -> FromFieldResult l a
+data Lenient =
+  SomeStrict !Strict |
+  IEEENaN |
+  IEEEInfinity |
+  IEEETooSmall |
+  IEEETooBig |
+  TextCouldNotParse
+  deriving stock (Eq, Show)
 
-deriving stock instance (Eq a) => Eq (FromFieldResult l a)
+data DecodeError (e :: Type) =
+  DecodeError !e {-# UNPACK #-} !TyCon
+  deriving stock (Eq, Show)
 
-deriving stock instance (Show a) => Show (FromFieldResult l a)
+class FromFieldStrict (a :: Type) where
+  fromFieldStrict :: MySQLValue -> Either (DecodeError Strict) a
 
-deriving stock instance Functor (FromFieldResult l)
-
-newtype L (a :: Type) = L a
-  deriving newtype (Eq)
-  deriving stock (Show)
-
-class FromField (l :: Leniency) (a :: Type) | a -> l where
-  fromField :: MySQLValue -> FromFieldResult l a
-
-instance FromField 'Strict Bool where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLInt8 v -> StrictParse (zeroBits /= v)
-    MySQLInt8U v -> StrictParse (zeroBits /= v)
-    MySQLBit v -> StrictParse (zeroBits /= v)
+instance FromFieldStrict Bool where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLInt8 v -> Right (zeroBits /= v)
+    MySQLInt8U v -> Right (zeroBits /= v)
+    MySQLBit v -> Right (zeroBits /= v)
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Bool) where
-  {-# INLINABLE fromField #-}
-  fromField = relax . fmap L . fromField
-
-instance FromField 'Strict Int8 where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLInt8 v -> StrictParse v
+instance FromFieldStrict Int8 where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLInt8 v -> Right v
     MySQLDecimal v -> tryScientific v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Int8) where
-  {-# INLINABLE fromField #-}
-  fromField v = fmap L $ case fromField v of
-    TypeMismatch' -> case v of
-      MySQLText v'   -> tryText v'
-      MySQLDouble v' -> tryIEEE v'
-      MySQLFloat v'  -> tryIEEE v'
-      _              -> TypeMismatch'
-    res          -> relax res
-
-instance FromField 'Strict Int16 where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLInt8 v -> StrictParse . fromIntegral $ v
-    MySQLInt16 v -> StrictParse v
+instance FromFieldStrict Int16 where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLInt8 v -> Right . fromIntegral $ v
+    MySQLInt16 v -> Right v
     MySQLDecimal v -> tryScientific v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Int16) where
-  {-# INLINABLE fromField #-}
-  fromField v = fmap L $ case fromField v of
-    TypeMismatch' -> case v of
-      MySQLText v'   -> tryText v'
-      MySQLDouble v' -> tryIEEE v'
-      MySQLFloat v'  -> tryIEEE v'
-      _              -> TypeMismatch'
-    res -> relax res
-
-instance FromField 'Strict Int32 where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLInt8 v -> StrictParse . fromIntegral $ v
-    MySQLInt16 v -> StrictParse . fromIntegral $ v
-    MySQLInt32 v -> StrictParse v
+instance FromFieldStrict Int32 where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLInt8 v -> Right . fromIntegral $ v
+    MySQLInt16 v -> Right . fromIntegral $ v
+    MySQLInt32 v -> Right v
     MySQLDecimal v -> tryScientific v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Int32) where
-  {-# INLINABLE fromField #-}
-  fromField v = fmap L $ case fromField v of
-    TypeMismatch' -> case v of
-      MySQLText v'   -> tryText v'
-      MySQLDouble v' -> tryIEEE v'
-      MySQLFloat v'  -> tryIEEE v'
-      _              -> TypeMismatch'
-    res -> relax res
-
-instance FromField 'Strict Int64 where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLInt8 v -> StrictParse . fromIntegral $ v
-    MySQLInt16 v -> StrictParse . fromIntegral $ v
-    MySQLInt32 v -> StrictParse . fromIntegral $ v
-    MySQLInt64 v -> StrictParse v
+instance FromFieldStrict Int64 where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLInt8 v -> Right . fromIntegral $ v
+    MySQLInt16 v -> Right . fromIntegral $ v
+    MySQLInt32 v -> Right . fromIntegral $ v
+    MySQLInt64 v -> Right v
     MySQLDecimal v -> tryScientific v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Int64) where
-  {-# INLINABLE fromField #-}
-  fromField v = fmap L $ case fromField v of
-    TypeMismatch' -> case v of
-      MySQLText v'   -> tryText v'
-      MySQLDouble v' -> tryIEEE v'
-      MySQLFloat v'  -> tryIEEE v'
-      _              -> TypeMismatch'
-    res -> relax res
-
-instance FromField 'Strict Int where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLInt8 v -> StrictParse . fromIntegral $ v
-    MySQLInt16 v -> StrictParse . fromIntegral $ v
+instance FromFieldStrict Int where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLInt8 v -> Right . fromIntegral $ v
+    MySQLInt16 v -> Right . fromIntegral $ v
     MySQLInt32 v -> tryFixed v
     MySQLInt64 v -> tryFixed v
     MySQLDecimal v -> tryScientific v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Int) where
-  {-# INLINABLE fromField #-}
-  fromField v = fmap L $ case fromField v of
-    TypeMismatch' -> case v of
-      MySQLText v'   -> tryText v'
-      MySQLDouble v' -> tryIEEE v'
-      MySQLFloat v'  -> tryIEEE v'
-      _              -> TypeMismatch'
-    res -> relax res
-
-instance FromField 'Strict Word8 where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLInt8U v -> StrictParse v
+instance FromFieldStrict Word8 where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLInt8U v -> Right v
     MySQLDecimal v -> tryScientific v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Word8) where
-  {-# INLINABLE fromField #-}
-  fromField = fmap L . relax . fromField
-
-instance FromField 'Strict Word16 where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLInt8U v -> StrictParse . fromIntegral $ v
-    MySQLInt16U v -> StrictParse v
+instance FromFieldStrict Word16 where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLInt8U v -> Right . fromIntegral $ v
+    MySQLInt16U v -> Right v
     MySQLDecimal v -> tryScientific v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Word16) where
-  {-# INLINABLE fromField #-}
-  fromField = fmap L . relax . fromField
-
-instance FromField 'Strict Word32 where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLInt8U v -> StrictParse . fromIntegral $ v
-    MySQLInt16U v -> StrictParse . fromIntegral $ v
-    MySQLInt32U v -> StrictParse v
+instance FromFieldStrict Word32 where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLInt8U v -> Right . fromIntegral $ v
+    MySQLInt16U v -> Right . fromIntegral $ v
+    MySQLInt32U v -> Right v
     MySQLDecimal v -> tryScientific v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Word32) where
-  {-# INLINABLE fromField #-}
-  fromField = fmap L . relax . fromField
-
-instance FromField 'Strict Word64 where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLInt8U v -> StrictParse . fromIntegral $ v
-    MySQLInt16U v -> StrictParse . fromIntegral $ v
-    MySQLInt32U v -> StrictParse . fromIntegral $ v
-    MySQLInt64U v -> StrictParse v
+instance FromFieldStrict Word64 where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLInt8U v -> Right . fromIntegral $ v
+    MySQLInt16U v -> Right . fromIntegral $ v
+    MySQLInt32U v -> Right . fromIntegral $ v
+    MySQLInt64U v -> Right v
     MySQLDecimal v -> tryScientific v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Word64) where
-  {-# INLINABLE fromField #-}
-  fromField = fmap L . relax . fromField
-
-instance FromField 'Strict Word where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLInt8U v -> StrictParse . fromIntegral $ v
-    MySQLInt16U v -> StrictParse . fromIntegral $ v
+instance FromFieldStrict Word where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLInt8U v -> Right . fromIntegral $ v
+    MySQLInt16U v -> Right . fromIntegral $ v
     MySQLInt32U v -> tryFixed v
     MySQLInt64U v -> tryFixed v
     MySQLDecimal v -> tryScientific v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Word) where
-  {-# INLINABLE fromField #-}
-  fromField = fmap L . relax . fromField
-
-instance FromField 'Strict Float where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLFloat v -> StrictParse v
+instance FromFieldStrict Float where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLFloat v -> Right v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Float) where
-  {-# INLINABLE fromField #-}
-  fromField v = fmap L $ case fromField v of
-    TypeMismatch' -> case v of
-      MySQLText v' -> tryIEEEFromText v'
-      _            -> TypeMismatch'
-    res          -> relax res
-
-instance FromField 'Strict Double where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLFloat v -> StrictParse . realToFrac $ v
-    MySQLDouble v -> StrictParse v
+instance FromFieldStrict Double where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLFloat v -> Right . realToFrac $ v
+    MySQLDouble v -> Right v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Double) where
-  {-# INLINABLE fromField #-}
-  fromField v = fmap L $ case fromField v of
-    TypeMismatch' -> case v of
-      MySQLText v' -> tryIEEEFromText v'
-      _            -> TypeMismatch'
-    res -> relax res
+instance FromFieldStrict SqlNull where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLNull -> Right SqlNull
+    _ -> Left . DecodeError TypeMismatch . typeRepTyCon $ (typeRep @SqlNull)
 
-instance FromField 'Strict Scientific where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLDecimal v -> StrictParse v
-    MySQLInt8 v -> StrictParse . fromIntegral $ v
-    MySQLInt8U v -> StrictParse . fromIntegral $ v
-    MySQLInt16 v -> StrictParse . fromIntegral $ v
-    MySQLInt16U v -> StrictParse . fromIntegral $ v
-    MySQLInt32 v -> StrictParse . fromIntegral $ v
-    MySQLInt32U v -> StrictParse . fromIntegral $ v
-    MySQLInt64 v -> StrictParse . fromIntegral $ v
-    MySQLInt64U v -> StrictParse . fromIntegral $ v
+instance FromFieldStrict ByteString where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLText v -> Right . encodeUtf8 $ v
+    MySQLBytes v -> Right v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Scientific) where
-  {-# INLINABLE fromField #-}
-  fromField = fmap L . relax . fromField
-
-instance FromField 'Strict Rational where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLInt8 v -> StrictParse . fromIntegral $ v
-    MySQLInt8U v -> StrictParse . fromIntegral $ v
-    MySQLInt16 v -> StrictParse . fromIntegral $ v
-    MySQLInt16U v -> StrictParse . fromIntegral $ v
-    MySQLInt32 v -> StrictParse . fromIntegral $ v
-    MySQLInt32U v -> StrictParse . fromIntegral $ v
-    MySQLInt64 v -> StrictParse . fromIntegral $ v
-    MySQLInt64U v -> StrictParse . fromIntegral $ v
+instance FromFieldStrict Text where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLText v -> Right v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Rational) where
-  {-# INLINABLE fromField #-}
-  fromField = fmap L . relax . fromField
-
-instance FromField 'Strict SqlNull where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLNull -> StrictParse SqlNull
-    _ -> TypeMismatch'
-
-instance FromField 'Lenient (L SqlNull) where
-  {-# INLINABLE fromField #-}
-  fromField = fmap L . relax . fromField
-
-instance FromField 'Strict ByteString where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLText v -> StrictParse . encodeUtf8 $ v
-    MySQLBytes v -> StrictParse v
+instance FromFieldStrict LocalTime where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLDateTime v -> Right v
+    MySQLTimeStamp v -> Right v
+    MySQLDate v -> Right . LocalTime v $ midnight
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L ByteString) where
-  {-# INLINABLE fromField #-}
-  fromField = fmap L . relax . fromField
-
-instance FromField 'Strict Text where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLText v -> StrictParse v
+instance FromFieldStrict Day where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
+    MySQLDate v -> Right v
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L Text) where
-  {-# INLINABLE fromField #-}
-  fromField v = fmap L $ case fromField v of
-    TypeMismatch' -> case v of
-      MySQLInt8 v'    -> LenientParse . pack . show $ v'
-      MySQLInt8U v'   -> LenientParse . pack . show $ v'
-      MySQLInt16 v'   -> LenientParse . pack . show $ v'
-      MySQLInt16U v'  -> LenientParse . pack . show $ v'
-      MySQLInt32 v'   -> LenientParse . pack . show $ v'
-      MySQLInt32U v'  -> LenientParse . pack . show $ v'
-      MySQLInt64 v'   -> LenientParse . pack . show $ v'
-      MySQLInt64U v'  -> LenientParse . pack . show $ v'
-      MySQLDecimal v' -> LenientParse . pack . show $ v'
-      MySQLFloat v'   -> LenientParse . pack . show $ v'
-      MySQLDouble v'  -> LenientParse . pack . show $ v'
-      _               -> TypeMismatch'
-    res          -> relax res
-
-instance FromField 'Strict LocalTime where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLDateTime v -> StrictParse v
-    MySQLTimeStamp v -> StrictParse v
-    MySQLDate v -> StrictParse . LocalTime v $ midnight
-    v -> handleNullOrMismatch v
-
-instance FromField 'Lenient (L LocalTime) where
-  {-# INLINABLE fromField #-}
-  fromField = fmap L . relax . fromField
-
-instance FromField 'Strict Day where
-  {-# INLINABLE fromField #-}
-  fromField = \case
-    MySQLDate v -> StrictParse v
-    v -> handleNullOrMismatch v
-
-instance FromField 'Lenient (L Day) where
-  {-# INLINABLE fromField #-}
-  fromField = fmap L . relax . fromField
-
-instance FromField 'Strict TimeOfDay where
-  {-# INLINABLE fromField #-}
-  fromField = \case
+instance FromFieldStrict TimeOfDay where
+  {-# INLINABLE fromFieldStrict #-}
+  fromFieldStrict = \case
     MySQLTime s v ->
       if s == zeroBits
-      then StrictParse v
-      else TypeMismatch'
+      then Right v
+      else
+        Left . DecodeError TypeMismatch . typeRepTyCon $ (typeRep @TimeOfDay)
     v -> handleNullOrMismatch v
 
-instance FromField 'Lenient (L TimeOfDay) where
-  {-# INLINABLE fromField #-}
-  fromField = fmap L . relax . fromField
+class (FromFieldStrict a) => FromFieldLenient (a :: Type) where
+  {-# INLINABLE fromFieldLenient #-}
+  fromFieldLenient :: MySQLValue -> Either (DecodeError Lenient) a
+  fromFieldLenient v = case fromFieldStrict v of
+    Left (DecodeError err t) -> Left . DecodeError (SomeStrict err) $ t
+    Right res                -> Right res
 
-data StrictDecodeError =
-  UnexpectedNull {-# UNPACK #-} !TyCon |
-  TypeMismatch {-# UNPACK #-} !TyCon |
-  Won'tFit {-# UNPACK #-} !TyCon
-  deriving stock (Eq, Show)
+instance (FromFieldStrict a,
+          Read a,
+          Typeable a,
+          Integral a,
+          Bounded a) =>
+  FromFieldLenient a where
+  {-# INLINABLE fromFieldLenient #-}
+  fromFieldLenient v = case fromFieldStrict v of
+    Left (DecodeError err t)    -> case err of
+      TypeMismatch -> case v of
+        MySQLText v'   -> tryText v'
+        MySQLFloat v'  -> tryIEEE v'
+        MySQLDouble v' -> tryIEEE v'
+        _              -> Left . DecodeError (SomeStrict err) $ t
+      _            -> Left . DecodeError (SomeStrict err) $ t
+    Right res -> Right res
 
-data LenientDecodeError =
-  SomeStrictError !StrictDecodeError |
-  IEEENaN {-# UNPACK #-} !TyCon |
-  IEEEInfinity {-# UNPACK #-} !TyCon |
-  IEEETooSmall {-# UNPACK #-} !TyCon |
-  IEEETooBig {-# UNPACK #-} !TyCon |
-  TextCouldNotParse {-# UNPACK #-} !TyCon
-  deriving stock (Eq, Show)
+instance FromFieldLenient Float where
+  {-# INLINABLE fromFieldLenient #-}
+  fromFieldLenient v = case fromFieldStrict v of
+    Left (DecodeError err t) -> case err of
+      TypeMismatch -> case v of
+        MySQLText v' -> tryIEEEFromText v'
+        _            -> Left . DecodeError (SomeStrict err) $ t
+      _ -> Left . DecodeError (SomeStrict err) $ t
+    Right res -> Right res
 
--- WARNING: Do _not_ write any instances of these classes! They are a trick to
--- avoid having to define everything 'beam' demands of us twice.
+instance FromFieldLenient Double where
+  {-# INLINABLE fromFieldLenient #-}
+  fromFieldLenient v = case fromFieldStrict v of
+    Left (DecodeError err t) -> case err of
+      TypeMismatch -> case v of
+        MySQLText v' -> tryIEEEFromText v'
+        _            -> Left . DecodeError (SomeStrict err) $ t
+      _ -> Left . DecodeError (SomeStrict err) $ t
+    Right res -> Right res
 
-class FromFieldS (a :: Type) where
-  fromFieldS :: MySQLValue -> Either StrictDecodeError a
-
-class FromFieldL (a :: Type) where
-  fromFieldL :: MySQLValue -> Either LenientDecodeError a
-
-instance (FromField 'Strict a, Typeable a) => FromFieldS a where
-  {-# INLINABLE fromFieldS #-}
-  fromFieldS v = case fromField v of
-    UnexpectedNull' -> Left . UnexpectedNull $ tyConOf @a
-    TypeMismatch'   -> Left . TypeMismatch $ tyConOf @a
-    Won'tFit'       -> Left . Won'tFit $ tyConOf @a
-    StrictParse v'  -> Right v'
-
-instance (FromField 'Lenient (L a), Typeable a) => FromFieldL a where
-  {-# INLINABLE fromFieldL #-}
-  fromFieldL v = case fromField v of
-    UnexpectedNull' -> Left . SomeStrictError . UnexpectedNull $ tyConOf @a
-    TypeMismatch' -> Left . SomeStrictError . TypeMismatch $ tyConOf @a
-    Won'tFit' -> Left . SomeStrictError . Won'tFit $ tyConOf @a
-    IEEENaN' -> Left . IEEENaN $ tyConOf @a
-    IEEEInfinity' -> Left . IEEEInfinity $ tyConOf @a
-    IEEETooSmall' -> Left . IEEETooSmall $ tyConOf @a
-    IEEETooBig' -> Left . IEEETooBig $ tyConOf @a
-    TextCouldNotParse' -> Left . TextCouldNotParse $ tyConOf @a
-    LenientParse (L v') -> Right v'
-    StrictParse (L v') -> Right v'
+instance FromFieldLenient Text where
+  {-# INLINABLE fromFieldLenient #-}
+  fromFieldLenient v = case fromFieldStrict v of
+    Left (DecodeError err t) -> case err of
+      TypeMismatch -> case v of
+        MySQLInt8 v'    -> Right . pack . show $ v'
+        MySQLInt8U v'   -> Right . pack . show $ v'
+        MySQLInt16 v'   -> Right . pack . show $ v'
+        MySQLInt16U v'  -> Right . pack . show $ v'
+        MySQLInt32 v'   -> Right . pack . show $ v'
+        MySQLInt32U v'  -> Right . pack . show $ v'
+        MySQLInt64 v'   -> Right . pack . show $ v'
+        MySQLInt64U v'  -> Right . pack . show $ v'
+        MySQLDecimal v' -> Right . pack . show $ v'
+        MySQLFloat v'   -> Right . pack . show $ v'
+        MySQLDouble v'  -> Right . pack . show $ v'
+        _               -> Left . DecodeError (SomeStrict err) $ t
+      _ -> Left . DecodeError (SomeStrict err) $ t
+    Right res -> Right res
 
 -- Helpers
 
-tyConOf :: forall (a :: Type) .
-  (Typeable a) => TyCon
-tyConOf = typeRepTyCon (typeRep @a)
+handleNullOrMismatch :: forall (a :: Type) .
+  (Typeable a) =>
+  MySQLValue -> Either (DecodeError Strict) a
+handleNullOrMismatch = Left . \case
+  MySQLNull -> DecodeError UnexpectedNull (tyCon @a)
+  _ -> DecodeError TypeMismatch (tyCon @a)
 
-typeName :: forall (a :: Type) .
-  (Typeable a) => Text
-typeName = pack . tyConName . typeRepTyCon $ (typeRep @a)
+tryScientific :: forall (a :: Type) .
+  (Typeable a, Integral a, Bounded a) =>
+  Scientific -> Either (DecodeError Strict) a
+tryScientific s = case toBoundedInteger s of
+  Nothing -> Left . DecodeError Won'tFit $ (tyCon @a)
+  Just v  -> Right v
 
-handleNullOrMismatch :: MySQLValue -> FromFieldResult l a
-handleNullOrMismatch = \case
-  MySQLNull -> UnexpectedNull'
-  _ -> TypeMismatch'
-
-relax :: FromFieldResult 'Strict a -> FromFieldResult 'Lenient a
-relax = \case
-  UnexpectedNull' -> UnexpectedNull'
-  TypeMismatch' -> TypeMismatch'
-  Won'tFit' -> Won'tFit'
-  StrictParse x -> StrictParse x
-
-tryScientific :: (Integral a, Bounded a) =>
-  Scientific -> FromFieldResult 'Strict a
-tryScientific = maybe Won'tFit' StrictParse . toBoundedInteger
-
-tryText :: (Read a) => Text -> FromFieldResult 'Lenient a
-tryText = maybe TextCouldNotParse' LenientParse . readMaybe . unpack
+tryText :: forall (a :: Type) .
+  (Read a, Typeable a) =>
+  Text -> Either (DecodeError Lenient) a
+tryText t = case readMaybe . unpack $ t of
+  Nothing -> Left . DecodeError TextCouldNotParse $ (tyCon @a)
+  Just x  -> Right x
 
 tryIEEE :: forall (b :: Type) (a :: Type) .
-  (RealFloat a, Integral b, Bounded b) =>
-  a -> FromFieldResult 'Lenient b
+  (Integral b, RealFloat a, Bounded b, Typeable b) =>
+  a -> Either (DecodeError Lenient) b
 tryIEEE v
-  | isNaN v = IEEENaN'
-  | isInfinite v = IEEEInfinity'
-  | v < fromIntegral (minBound @b) = IEEETooSmall'
-  | v > fromIntegral (maxBound @b) = IEEETooBig'
-  | otherwise = LenientParse . truncate $ v
+  | isNaN v = Left . DecodeError IEEENaN $ (tyCon @b)
+  | isInfinite v = Left . DecodeError IEEEInfinity $ (tyCon @b)
+  | v < fromIntegral (minBound @b) =
+      Left. DecodeError IEEETooSmall $ (tyCon @b)
+  | v > fromIntegral (maxBound @b) =
+      Left . DecodeError IEEETooBig $ (tyCon @b)
+  | otherwise = Right . truncate $ v
 
-tryFixed :: (Integral a, Integral b, Bits a, Bits b) =>
-  a -> FromFieldResult 'Strict b
-tryFixed = maybe Won'tFit' StrictParse . toIntegralSized
-
-tryIEEEFromText :: (Read a, RealFloat a) =>
-  Text -> FromFieldResult 'Lenient a
+tryIEEEFromText :: forall (a :: Type) .
+  (Read a, RealFloat a, Typeable a) =>
+  Text -> Either (DecodeError Lenient) a
 tryIEEEFromText = \case
-  "Infinity" -> LenientParse . read $ "Infinity"
-  "-Infinity" -> LenientParse . read $ "-Infinity"
+  "Infinity" -> Right . read $ "Infinity"
+  "-Infinity" -> Right . read $ "-Infinity"
   t -> case readMaybe . unpack $ t of
-    Nothing -> TextCouldNotParse'
+    Nothing -> Left . DecodeError TextCouldNotParse $ (tyCon @a)
     Just v -> if isInfinite v
-              then case signum v of
-                -1 -> IEEETooSmall'
-                _  -> IEEETooBig'
-              else LenientParse v
+                then case signum v of
+                  -1 -> Left . DecodeError IEEETooSmall $ (tyCon @a)
+                  _  -> Left . DecodeError IEEETooBig $ (tyCon @a)
+                else Right v
+
+tryFixed :: forall (b :: Type) (a :: Type) .
+  (Integral a, Integral b, Bits a, Bits b, Typeable b) =>
+  a -> Either (DecodeError Strict) b
+tryFixed v = case toIntegralSized v of
+  Nothing -> Left . DecodeError Won'tFit $ (tyCon @b)
+  Just v' -> Right v'
+
+tyCon :: forall (a :: Type) .
+  (Typeable a) =>
+  TyCon
+tyCon = typeRepTyCon (typeRep @a)